remove duplicate util file for redirect status
