+++ packages/next/src/client/components/redirect.ts
@@ -106,3 +106,13 @@ export function getRedirectTypeFromError<U extends string>(
 
   return error.digest.split(';', 2)[1] as RedirectType
 }
+
+export function getRedirectStatusCodeFromError<U extends string>(
+  error: RedirectError<U>
+): number {
+  if (!isRedirectError(error)) {
+    throw new Error('Not a redirect error')
+  }
+
+  return error.digest.split(';', 4)[3] === 'true' ? 308 : 307
+}
